skip old tests that need to be updated to the new data model
