+++ seed/data_importer/tests/test_views.py
@@ -1,0 +1,127 @@
+# !/usr/bin/env python
+# encoding: utf-8
+"""
+:copyright (c) 2014 - 2016, The Regents of the University of California, through Lawrence Berkeley National Laboratory (subject to receipt of any required approvals from the U.S. Department of Energy) and contributors. All rights reserved.  # NOQA
+:author
+"""
+import json
+
+from django.core.urlresolvers import reverse_lazy
+from django.test import TestCase
+
+from seed.data_importer.models import ROW_DELIMITER, ImportFile, ImportRecord
+from seed.landing.models import SEEDUser as User
+
+
+class DataImporterViewTests(TestCase):
+    """
+    Tests of the data_importer views (and the objects they create).
+    """
+
+    def setUp(self):
+        user_details = {
+            'username': 'test_user@demo.com',
+            'password': 'test_pass',
+        }
+        self.user = User.objects.create_superuser(
+            email='test_user@demo.com', **user_details)
+        self.client.login(**user_details)
+
+    def test_get_raw_column_names(self):
+        """Make sure we get column names back in a format we expect."""
+        import_record = ImportRecord.objects.create()
+        expected_raw_columns = ['tax id', 'name', 'etc.']
+        expected_saved_format = ROW_DELIMITER.join(expected_raw_columns)
+        import_file = ImportFile.objects.create(
+            import_record=import_record,
+            cached_first_row=expected_saved_format
+        )
+
+        # Just make sure we were saved correctly
+        self.assertEqual(import_file.cached_first_row, expected_saved_format)
+
+        url = reverse_lazy("seed:get_raw_column_names")
+        resp = self.client.post(
+            url, data=json.dumps(
+                {'import_file_id': import_file.pk}
+            ), content_type='application/json'
+        )
+
+        body = json.loads(resp.content)
+
+        self.assertEqual(body.get('raw_columns', []), expected_raw_columns)
+
+    def test_get_first_five_rows(self):
+        """Make sure we get our first five rows back correctly."""
+        import_record = ImportRecord.objects.create()
+        expected_raw_columns = ['tax id', 'name', 'etc.']
+        expected_raw_rows = [
+            ['02023', '12 Jefferson St.', 'etc.'],
+            ['12433', '23 Washington St.', 'etc.'],
+            ['04422', '4 Adams St.', 'etc.'],
+        ]
+
+        expected = [
+            dict(zip(expected_raw_columns, row)) for row in expected_raw_rows
+            ]
+        expected_saved_format = '\n'.join([
+                                              ROW_DELIMITER.join(row) for row
+                                              in expected_raw_rows
+                                              ])
+        import_file = ImportFile.objects.create(
+            import_record=import_record,
+            cached_first_row=ROW_DELIMITER.join(expected_raw_columns),
+            cached_second_to_fifth_row=expected_saved_format
+        )
+
+        # Just make sure we were saved correctly
+        self.assertEqual(
+            import_file.cached_second_to_fifth_row, expected_saved_format
+        )
+
+        url = reverse_lazy("seed:get_first_five_rows")
+        resp = self.client.post(
+            url, data=json.dumps(
+                {'import_file_id': import_file.pk}
+            ), content_type='application/json'
+        )
+
+        body = json.loads(resp.content)
+
+        self.assertEqual(body.get('first_five_rows', []), expected)
+
+    def test_get_first_five_rows_with_newlines(self):
+        import_record = ImportRecord.objects.create()
+        expected_raw_columns = ['id', 'name', 'etc']
+        expected_raw_rows = [
+            ['1', 'test', 'new\nline'],
+            ['2', 'test', 'single'],
+        ]
+
+        expected = [
+            dict(zip(expected_raw_columns, row)) for row in expected_raw_rows
+            ]
+        expected_saved_format = "1%stest%snew\nline\n2%stest%ssingle".replace(
+            '%s', ROW_DELIMITER)
+
+        import_file = ImportFile.objects.create(
+            import_record=import_record,
+            cached_first_row=ROW_DELIMITER.join(expected_raw_columns),
+            cached_second_to_fifth_row=expected_saved_format
+        )
+
+        # Just make sure we were saved correctly
+        self.assertEqual(
+            import_file.cached_second_to_fifth_row, expected_saved_format
+        )
+
+        url = reverse_lazy("seed:get_first_five_rows")
+        resp = self.client.post(
+            url, data=json.dumps(
+                {'import_file_id': import_file.pk}
+            ), content_type='application/json'
+        )
+
+        body = json.loads(resp.content)
+
+        self.assertEqual(body.get('first_five_rows', []), expected)
